Finally, KSCD uses newer WorkMan code. This means, that KSCD now needs libWorkMan (included in libwm/) to work. Because the code (both libworkman and kscd) is a little rough cut, I await lots of problems running kscd now. At least on platforms other than Linux and Solaris.
I hope to fix this ASAP, especially after removing the compile time
configuration.

Be prepared to see lots of output on stderr with this version of kscd.
This comes from the libWorkMan diagnostics and is run-time configurable
(but I didn't implement it in kscd so far).

Please forgive me my incomplete work in the makefiles, but I've got no
time left and had to commit.

svn path=/trunk/kdemultimedia/kscd/; revision=51207