Fix changing audio backends in digital mode again.
With my previous fix changing the property from currentItem
to currentText, the current backend was set with QComboBox::setCurrentText,
which replaces the text in the current item if the text isn't in the list.
So because the default backend is an empty string, "arts" was replaced by ""
and it wasn't possible to change the backend to "arts".

Instead, make the combo box a sub-class of KComboBox where setCurrentText
calls KComboBox::setCurrentItem(text, false) so the text isn't changed if
it isn't in the list.

Also, remove the default audio-device again because arts doesn't use the
device and the alsa devices are in a different format


svn path=/branches/KDE/3.5/kdemultimedia/kscd/; revision=488028